Update package names, unit tests paths
