+++ src/pluto/common/media/TextDetector.py
