+++ src/pluto/ui/qt/mvc/routers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+
 class Router(object):
     def __init__(self, app):
         self.app = app
